Account for move of Serializable to dmod.core.
