+++ python/services/monitorservice/setup.py
@@ -14,7 +14,7 @@ setup(
     author_email='',
     url='',
     license='',
-    install_requires=['dmod-communication>=0.4.2', 'dmod-monitor>=0.3.0'],
+    install_requires=['dmod-core>=0.1.0', 'dmod-communication>=0.4.2', 'dmod-monitor>=0.3.0'],
     packages=find_namespace_packages(exclude=('tests', 'schemas', 'ssl', 'src'))
 )
