+++ python/services/requestservice/setup.py
@@ -14,7 +14,7 @@ setup(
     author_email='',
     url='',
     license='',
-    install_requires=['websockets', 'dmod-communication>=0.5.0', 'dmod-access>=0.2.0',
+    install_requires=['websockets', 'dmod-core>=0.1.0', 'dmod-communication>=0.5.0', 'dmod-access>=0.2.0',
                       'dmod-externalrequests>=0.1.1'],
     packages=find_namespace_packages(exclude=('tests', 'schemas', 'ssl', 'src'))
 )
